pyproject.toml
[project]
name = "undercover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "httpx[socks]",
    "langgraph",
    "langgraph-cli[inmem]",
    "langchain-tavily",
    "python-dotenv",
    "pytest",
    "pydantic",
    "PyYAML",
    "black",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "starlette>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
